pyproject.toml
[project]
name = "msg-to-event"
version = "0.1.0"
description = ""
authors = [
    {name = "eivasch"}
]
readme = "README.md"
requires-python = ">=3.13"
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python-telegram-bot = "^22.0"
dotenv = "^0.9.9"
openai = "^1.76.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
mypy = "^1.15.0"

[tool.ruff.lint]
extend-select = [
  "W292",  # missing-newline-at-end-of-file
  "I", # isort
]
